requirements.txt
asn1crypto
atomicwrites
attrs
certifi
cffi
chardet
colorama
cryptography
idna
inflection
lz4
more-itertools
Naked
pluggy
py
pycparser
pycryptodome
pytest
pytest-asyncio
python-lzf
PyYAML
requests
shellescape
six
urllib3
websockets
